# Remove the use of the gettimeofday function