Test 1 : see ~/caja-debug-log.txt
```
#!/bin/bash
cat << EOF > ~/caja-debug-log.conf
[debug log]
max lines = 1000
enable domains = async;GLog
EOF

# To write the debug log to disk immediately, send a SIGUSR1 to Caja:
# kill -s SIGUSR1 $(pgrep caja)

# To see Caja log:
# cat ~/caja-debug-log.txt
```
Test 2 : call ./test/test-eel-pixbuf-scale FILENAME
```
./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png
```