@raveit65 2020-11-15
Test 2 seems OK, btw. it scales faster now. (msecs --> usecs)
master:
```
[rave@mother caja-test]$ ./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png
scale factors: 16,281250, 5,125000
Time for eel_gdk_pixbuf_scale_down: 10 msecs
Time for INTERP_NEAREST: 0 msecs
Time for INTERP_BILINEAR: 21 msecs
[rave@mother caja-test]$ ./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png
scale factors: 16,281250, 5,125000
Time for eel_gdk_pixbuf_scale_down: 9 msecs
Time for INTERP_NEAREST: 0 msecs
Time for INTERP_BILINEAR: 20 msecs
[rave@mother caja-test]$ ./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png
scale factors: 16,281250, 5,125000
Time for eel_gdk_pixbuf_scale_down: 15 msecs
Time for INTERP_NEAREST: 0 msecs
Time for INTERP_BILINEAR: 21 msecs
```
with that branch:
```
[rave@mother caja-dist]$ ./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png 
scale factors: 16,281250, 5,125000
Time for eel_gdk_pixbuf_scale_down: 150 usecs
Time for INTERP_NEAREST: 14 usecs
Time for INTERP_BILINEAR: 373 usecs
[rave@mother caja-dist]$ ./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png 
scale factors: 16,281250, 5,125000
Time for eel_gdk_pixbuf_scale_down: 107 usecs
Time for INTERP_NEAREST: 8 usecs
Time for INTERP_BILINEAR: 268 usecs
[rave@mother caja-dist]$ ./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png 
scale factors: 16,281250, 5,125000
Time for eel_gdk_pixbuf_scale_down: 159 usecs
Time for INTERP_NEAREST: 10 usecs
Time for INTERP_BILINEAR: 308 usecs
[rave@mother caja-dist]$ ./test/test-eel-pixbuf-scale /usr/share/pixmaps/fedora-logo.png 
scale factors: 16,281250, 5,125000
Time for eel_gdk_pixbuf_scale_down: 258 usecs
Time for INTERP_NEAREST: 24 usecs
Time for INTERP_BILINEAR: 656 usecs
```

But i am missing the the date in `caja-debug-log.txt` with test 1.
```
[rave@mother ~]$ cat caja-debug-log.txt.old 
===== BEGIN MILESTONES =====
0x562b12d2de90 2020/11/15 11:29:04.7438 (USER): user requested dump of debug log
===== END MILESTONES =====
===== BEGIN RING BUFFER =====
0x562b12d2de90 2020/11/15 11:29:01.3358 (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x562b12d2de90 2020/11/15 11:29:01.3474 (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x562b12d2de90 2020/11/15 11:29:01.3488 (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x562b12d2de90 2020/11/15 11:29:01.4140 (USER): window 0x562b130da4c0 open location: old="(none)", new="x-caja-desktop:///"
0x562b12d2de90 2020/11/15 11:29:01.4236 (USER): Bus acquired at org.freedesktop.FileManager1
0x562b12d2de90 2020/11/15 11:29:01.4242 (USER): Acquired the name org.freedesktop.FileManager1 on the session message bus

0x562b12d2de90 2020/11/15 11:29:02.1275 (USER): finished loading window 0x562b130da4c0: x-caja-desktop:///
0x562b12d2de90 2020/11/15 11:29:04.7438 (USER): user requested dump of debug log
===== END RING BUFFER =====


This configuration for the debug log can be re-created
by putting the following in ~/caja-debug-log.conf
(use ';' to separate domain names):

[debug log]
max lines=1000
[rave@mother ~]$ 
```
with that branch:
```
[rave@mother ~]$ cat caja-debug-log.txt
===== BEGIN MILESTONES =====
0x55ce9fd6ee90 (null) (USER): user requested dump of debug log
===== END MILESTONES =====
===== BEGIN RING BUFFER =====
0x55ce9fd6ee90 (null) (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x55ce9fd6ee90 (null) (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x55ce9fd6ee90 (null) (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x55ce9fd6ee90 (null) (USER): window 0x55cea011c500 open location: old="(none)", new="x-caja-desktop:///"
0x55ce9fd6ee90 (null) (USER): Bus acquired at org.freedesktop.FileManager1
0x55ce9fd6ee90 (null) (USER): Acquired the name org.freedesktop.FileManager1 on the session message bus

0x55ce9fd6ee90 (null) (USER): finished loading window 0x55cea011c500: x-caja-desktop:///
0x55ce9fd6ee90 (null) (USER): create new navigation window=0x55cea07623d0
0x55ce9fd6ee90 (null) (USER): window 0x55cea07623d0 open location: old="(none)", new="file:///home/rave"
0x55ce9fd6ee90 (null) (USER): change view of window 0x55cea07623d0: "file:///home/rave" to "OAFIID:Caja_File_Manager_Icon_View"
0x55ce9fd6ee90 (null) (USER): finished loading window 0x55cea07623d0: file:///home/rave
0x55ce9fd6ee90 (null) (USER): user requested dump of debug log
===== END RING BUFFER =====


This configuration for the debug log can be re-created
by putting the following in ~/caja-debug-log.conf
(use ';' to separate domain names):

[debug log]
max lines=1000
```
Is that intend?