@raveit65 2020-11-15
> .....but I do not know how to fully test this in terms of the caja-debug-log file's function unless this is it.

1. create `~/caja-debug-log.conf `
```
[rave@mother ~]$ cat ~/caja-debug-log.conf 
#!/bin/bash
cat << EOF > ~/caja-debug-log.conf
[debug log]
max lines = 1000
enable domains = async;GLog
EOF

# To write the debug log to disk immediately, send a SIGUSR1 to Caja:
# kill -s SIGUSR1 $(pgrep caja)

# To see Caja log:
# cat ~/caja-debug-log.txt
```
2. make it executable
3. run `kill -s SIGUSR1 $(pgrep caja)`

This will create `~/caja-debug-log.txt`

For some reasons a `killall caja` is needed to create `~/caja-debug-log.txt` again.