@raveit65 2020-11-20
@rbuj 
Before we missed this PR.
The date is missing in caja-debug-log.txt with this PR. See my previous comment.