@lukefromdc 2020-11-20
What is the purpose removing gettimeofday?