@rbuj 2020-11-21
There is no Y2038 problem: nowadays, time_t is 64-bit on almost all platforms.
```C
       #include <sys/time.h>

       int gettimeofday(struct timeval *tv, struct timezone *tz);

           struct timeval {
               time_t      tv_sec;     /* seconds */
               suseconds_t tv_usec;    /* microseconds */
           };
```
However, if we use glib, we can consider gettimeofday a low-level API, which in this case has other better choices:
- g_date_time_format: It's a function intended to use for printing dates with specific placeholders
- g_get_monotonic_time: it's a more suitable function for measuring short time intervals