@raveit65 2020-11-23
@rbuj 
once again, the date is missing in log with this PR.
```
[rave@mother ~]$ cat caja-debug-log.txt
===== BEGIN MILESTONES =====
0x55ce9fd6ee90 (null) (USER): user requested dump of debug log
===== END MILESTONES =====
===== BEGIN RING BUFFER =====
0x55ce9fd6ee90 (null) (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x55ce9fd6ee90 (null) (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x55ce9fd6ee90 (null) (USER): Error loading keys from file: Datei oder Verzeichnis nicht gefunden

0x55ce9fd6ee90 (null) (USER): window 0x55cea011c500 open location: old="(none)", new="x-caja-desktop:///"
0x55ce9fd6ee90 (null) (USER): Bus acquired at org.freedesktop.FileManager1
0x55ce9fd6ee90 (null) (USER): Acquired the name org.freedesktop.FileManager1 on the session message bus

0x55ce9fd6ee90 (null) (USER): finished loading window 0x55cea011c500: x-caja-desktop:///
0x55ce9fd6ee90 (null) (USER): create new navigation window=0x55cea07623d0
0x55ce9fd6ee90 (null) (USER): window 0x55cea07623d0 open location: old="(none)", new="file:///home/rave"
0x55ce9fd6ee90 (null) (USER): change view of window 0x55cea07623d0: "file:///home/rave" to "OAFIID:Caja_File_Manager_Icon_View"
0x55ce9fd6ee90 (null) (USER): finished loading window 0x55cea07623d0: file:///home/rave
0x55ce9fd6ee90 (null) (USER): user requested dump of debug log
===== END RING BUFFER =====
```